feat: fix the in-composer AI layers rendering behind the docked compose window by raising the portaled sparkle/pill/menu z-indexes above window chrome (still under the confirm layer), give the compose header a proper identity (sky pen chip plus two-line title with the recipient or draft state), add the Template picker and booking-link tools to the compose action bar by extracting TemplatePickerContent and InsertBookingLink out of ReplyComposer into shared unibox components, and make Compose the house primary sky button in the rail and mobile header
