+++ web/src/components/app/unibox/UniboxHeader.tsx
@@ -99,9 +99,9 @@ export function UniboxHeader({
                     <button
                         type="button"
                         onClick={() => useComposeStore.getState().openCompose()}
-                        className="lg:hidden inline-flex items-center gap-1.5 h-6 px-2.5 rounded-lg bg-white border border-slate-200 hover:bg-slate-50 hover:border-slate-300 text-slate-700 hover:text-slate-900 text-[11.5px] font-medium shadow-sm transition-colors"
+                        className="lg:hidden inline-flex items-center gap-1.5 h-6 px-2.5 rounded-lg bg-sky-600 hover:bg-sky-700 text-white text-[11.5px] font-medium shadow-sm shadow-sky-600/20 transition-colors"
                     >
-                        <PenLineIcon className="w-3 h-3 text-sky-600" />
+                        <PenLineIcon className="w-3 h-3" />
                         Compose
                     </button>
                 </ShortcutTooltip>
